Add files via upload
Bug fixes:
- Fixed a buffer overflow in cases where the amount of data input exceeds 2^32.
- Fixed a bug that kept the turbine output always at the inflow level.

Added a file ("plot_waterlevel.py") to plot the water level, turbine level and overflow over time. The plots present the the calculated data in a visual way and help to find inconsistencies in the logic.
Added two variables in the config file that allow the restrict the output in the timei file to a certain time frame.